New picture for quickstart in rviz tutorial. Please read the rest of the commit message.
In "Step 3:Interact with the Panda" the instrutions indicate you should check and uncheck some boxes for visualizations, but then the related image does not have those same boxes checked.

This could lead to confustion for a new user as the provided image shows the screne robot but if they checked the boxes correctly, they will not have a scene robot in their rviz.

This commit has an image with the correct boxes checked, but my rviz is slightly different (arrows instead of checkboxes, no rivets on the Panda, etc), so I wont be offened if you just make your own image to replace screen4.png